pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-ssg"
version = "0.1.0"
authors = [
    {name = "Brady Clarke", email = "hello@bradyclarke.com"},
]
description = "A minimalist static site generator"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
]
dependencies = [
    "markdown>=3.3.0",
    "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/bradyclarke/simple-ssg"
"Bug Tracker" = "https://github.com/bradyclarke/simple-ssg/issues"

[project.scripts]
simple-ssg = "simple_ssg.cli:main"

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "B", "I", "W", "C4", "UP", "N", "ANN", "DTZ", "COM"]
ignore = ["ANN101", "ANN401"]

[tool.ruff.isort]
known-first-party = ["simple_ssg"]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ANN", "E501"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
